refactor: web/ext -> ext
