+++ Makefile
@@ -3,7 +3,7 @@ default: test lint
 test:
 	go test ./...
 	cd web/webgen/gen-chi/webruntime && go test ./...
-	cd web/ext/scroll/internal && go test ./...
+	cd ext/scroll/internal && go test ./...
 .PHONY: test
 
 lint:
