+++ web/webgen/gen-chi/generator.go
@@ -15,7 +15,7 @@ import (
 	"github.com/podhmo/apikit/pkg/tinypkg"
 	"github.com/podhmo/apikit/resolve"
 	"github.com/podhmo/apikit/web"
-	"github.com/podhmo/apikit/web/ext"
+	"github.com/podhmo/apikit/ext"
 )
 
 type Config struct {
